feat(nip42,nip70): add auth and protected events.
