--- deployment/dev/docker-compose.yml
@@ -1,10 +1,0 @@
-version: '3.8'
-
-services:
-  mongo:
-    image: mongo:latest
-    environment:
-      MONGO_INITDB_ROOT_USERNAME: mongo_user
-      MONGO_INITDB_ROOT_PASSWORD: mongo_password
-    ports:
-      - "27017:27017"
