+++ documents/NIPs.md
@@ -7,11 +7,10 @@ The Immortal follows [NIPs](https://github.com/nostr-protocol/nips) and tries to
 - [X] **NIP-11**: Relay Information Document
 - [X] **NIP-13**: Proof of Work
 - [ ] **NIP-40**: Expiration Timestamp
-- [ ] **NIP-42**: Authentication of Clients to Relays
+- [X] **NIP-42**: Authentication of Clients to Relays
 - [ ] **NIP-50**: Search Capability
 - [ ] **NIP-56**: Reporting
-- [ ] **NIP-59**: Gift Wrap
-- [ ] **NIP-70**: Protected Events
+- [X] **NIP-70**: Protected Events
 - [ ] **NIP-94**: File Metadata
 - [ ] **NIP-96**: HTTP File Storage Integration
 - [ ] **NIP-98**: HTTP Auth
