+++ types/nip11/nip11.go
@@ -17,6 +17,7 @@ type RelayInformationDocument struct {
 	PaymentsURL    string                   `json:"payments_url,omitempty"`
 	Fees           *RelayFeesDocument       `json:"fees,omitempty"`
 	Icon           string                   `json:"icon"`
+	URL            string                   `json:"-"`
 }
 
 type RelayLimitationDocument struct {
